requirements.txt
slack_sdk
python-dotenv
openai
pyyaml
